프로필 이미지 처리 완료. 어드민 회원 수정 남음.

+++ src/main/java/com/utime/burrowNest/admin/vo/ManageUserVo.java
@@ -15,7 +15,7 @@ public class ManageUserVo extends UserVo {
 	/**
 	 * 총 로그인 횟수
 	 */
-	long loginCount;
+	int loginCount;
 	/** 마지막 로그인 날짜 */
 	LocalDateTime lastLoginTime;
 	/** 로그인 실패 횟수 */
